+++ docs/internal/processes/development-guide-relevant-files.md
@@ -216,6 +216,7 @@ This inventory records the checked-in files and directories that the maintainer 
 - Keep canonical detached-copy ownership for runtime tokens, token histories, token colors, provider-session metadata, and provider-failure metadata in `pkg/interfaces/world_state_clones.go`; backend packages that need those shapes should call the shared helpers instead of maintaining package-local clone trees.
 - When runtime snapshot, transitioner, token-transformer, worker, or replay code needs detached copies of canonical runtime tokens or provider metadata, route those copies through `pkg/interfaces/world_state_clones.go` helpers such as `interfaces.CloneToken(...)`, `interfaces.CloneTokens(...)`, `interfaces.CloneTokenColor(...)`, `interfaces.CloneTokenHistory(...)`, `interfaces.CloneProviderSessionMetadata(...)`, and `interfaces.CloneProviderFailureMetadata(...)` instead of reintroducing package-local clones in `pkg/factory/engine`, `pkg/factory/subsystems`, `pkg/workers`, or `pkg/replay`.
 - When backend size cleanup touches runtime orchestration packages such as `pkg/factory/runtime`, `pkg/factory/engine`, `pkg/factory/scheduler`, `pkg/factory/subsystems`, `pkg/replay`, or `pkg/config`, prefer splitting long constructors or execution paths into same-package phase helpers that keep the public seam and dependency direction unchanged instead of introducing a new helper package just to satisfy the file or function limit.
+- When backend size cleanup touches schema-heavy config or projection packages such as `pkg/config` or `pkg/factory/projections`, prefer same-package sibling files that separate decode, validation, expansion, dispatch, or generated-conversion helpers so the public seam stays stable while file-size ownership becomes reviewable by concern.
 - When `pkg/service.FactoryService` is handed to `APIServerStarter` as an `apisurface.APISurface`, prove seam-preservation through the callback-owned `runtime` parameter by exercising submit, subscribe, snapshot, or current-factory behavior there instead of adding reflection or interface-inventory assertions.
 - When runtime-state cleanup changes token or place bookkeeping, prove the contract through caller-facing mutation or reconstruction tests that assert owned state surfaces such as `Marking.PlaceTokens` or reconstructed `PlaceOccupancyByID`; do not add helper-topology or source-layout assertions.
 - When the CLI docs command surface changes, update `README.md` in the same pass and keep its command examples aligned with the guard assertions in `pkg/cli/root_test.go`.
