+++ pkg/factory/projections/world_state.go
@@ -7,7 +7,6 @@ import (
 
 	factoryapi "github.com/portpowered/infinite-you/pkg/api/generated"
 	"github.com/portpowered/infinite-you/pkg/interfaces"
-	"github.com/portpowered/infinite-you/pkg/workcontent"
 )
 
 const (
@@ -230,504 +229,6 @@ func (r *factoryWorldReducer) applyWorkRequest(context factoryapi.FactoryEventCo
 
 func (r *factoryWorldReducer) applyRelationshipChange(context factoryapi.FactoryEventContext, payload factoryapi.RelationshipChangeRequestEventPayload) {
 	r.addRelation(r.factoryRelationFromGenerated(payload.Relation, context))
-}
-
-func (r *factoryWorldReducer) applyDispatchCreated(event factoryapi.FactoryEvent, payload factoryapi.DispatchRequestEventPayload) {
-	dispatchID := stringValue(event.Context.DispatchId)
-	if dispatchID == "" {
-		return
-	}
-	inputWorkIDs := dispatchInputWorkIDs(payload, event.Context.WorkIds)
-	workIDs := make([]string, 0, len(inputWorkIDs))
-	traceIDs := make([]string, 0, len(inputWorkIDs))
-	inputWorkItems := make([]interfaces.FactoryWorkItem, 0, len(inputWorkIDs))
-	inputs := make([]interfaces.WorkstationInput, 0, len(inputWorkIDs))
-	for _, workID := range inputWorkIDs {
-		if workID == "" {
-			continue
-		}
-		item, ok := r.stateValue.WorkItemsByID[workID]
-		if !ok {
-			item = interfaces.FactoryWorkItem{ID: workID}
-		}
-		if item.TraceID == "" {
-			item.TraceID = firstString(event.Context.TraceIds)
-		}
-		placeID := r.workPlaces[item.ID]
-		if placeID == "" {
-			placeID = item.PlaceID
-		}
-		if placeID == "" {
-			placeID = r.initialPlaceForWorkType(item.WorkTypeID)
-		}
-		item.PlaceID = placeID
-		r.removeToken(item.ID)
-		r.stateValue.WorkItemsByID[item.ID] = item
-		r.stateValue.ActiveWorkItemsByID[item.ID] = item
-		workIDs = appendUnique(workIDs, item.ID)
-		traceIDs = appendUnique(traceIDs, item.TraceID)
-		r.addTraceWork(item.TraceID, item.ID)
-		inputWorkItems = append(inputWorkItems, item)
-		inputs = append(inputs, interfaces.WorkstationInput{
-			TokenID:  item.ID,
-			PlaceID:  placeID,
-			WorkItem: &item,
-		})
-	}
-
-	worker := r.workerForTransition(payload.TransitionId)
-	runnerID, runnerSource := generatedDispatchRunnerSelection(payload.Metadata)
-	dispatch := interfaces.FactoryWorldDispatch{
-		DispatchID:            dispatchID,
-		TransitionID:          payload.TransitionId,
-		Workstation:           r.workstationRefForTransition(payload.TransitionId),
-		RunnerID:              runnerID,
-		RunnerSelectionSource: runnerSource,
-		Provider:              worker.Provider,
-		Model:                 worker.Model,
-		StartedTick:           event.Context.Tick,
-		StartedAt:             event.Context.EventTime,
-		Inputs:                inputs,
-		WorkItemIDs:           sortedStrings(workIDs),
-		CurrentChainingTraceID: dispatchCurrentChainingTraceID(
-			event.Context.CurrentChainingTraceId,
-			payload.CurrentChainingTraceId,
-			inputWorkItems,
-		),
-		PreviousChainingTraceIDs: dispatchPreviousChainingTraceIDs(
-			event.Context.PreviousChainingTraceIds,
-			payload.PreviousChainingTraceIds,
-			inputWorkItems,
-		),
-		TraceIDs: interfaces.CanonicalChainingTraceIDs(traceIDs),
-	}
-	dispatch.Resources = r.consumeResourceUnits(payload.Resources)
-	r.stateValue.ActiveDispatches[dispatchID] = dispatch
-	for _, traceID := range dispatch.TraceIDs {
-		r.addTraceDispatch(traceID, dispatchID)
-	}
-}
-
-func dispatchInputWorkIDs(payload factoryapi.DispatchRequestEventPayload, contextWorkIDs *[]string) []string {
-	ordered := make([]string, 0, len(payload.Inputs)+len(sliceValue(contextWorkIDs)))
-	for _, ref := range payload.Inputs {
-		ordered = appendUnique(ordered, ref.WorkId)
-	}
-	for _, workID := range sliceValue(contextWorkIDs) {
-		ordered = appendUnique(ordered, workID)
-	}
-	return ordered
-}
-
-func (r *factoryWorldReducer) applyInferenceRequest(event factoryapi.FactoryEvent, payload factoryapi.InferenceRequestEventPayload) {
-	dispatchID := stringValue(event.Context.DispatchId)
-	if dispatchID == "" || payload.InferenceRequestId == "" {
-		return
-	}
-	attempts := r.inferenceAttemptsForDispatch(dispatchID)
-	current := attempts[payload.InferenceRequestId]
-	current.DispatchID = dispatchID
-	current.TransitionID = firstNonEmpty(current.TransitionID, r.transitionIDForDispatch(dispatchID))
-	current.InferenceRequestID = payload.InferenceRequestId
-	current.Attempt = payload.Attempt
-	current.WorkingDirectory = payload.WorkingDirectory
-	current.Worktree = payload.Worktree
-	current.Prompt = payload.Prompt
-	current.RequestTime = event.Context.EventTime
-	attempts[payload.InferenceRequestId] = current
-}
-
-func (r *factoryWorldReducer) applyInferenceResponse(event factoryapi.FactoryEvent, payload factoryapi.InferenceResponseEventPayload) {
-	dispatchID := stringValue(event.Context.DispatchId)
-	if dispatchID == "" || payload.InferenceRequestId == "" {
-		return
-	}
-	attempts := r.inferenceAttemptsForDispatch(dispatchID)
-	current := attempts[payload.InferenceRequestId]
-	current.DispatchID = dispatchID
-	current.TransitionID = firstNonEmpty(current.TransitionID, r.transitionIDForDispatch(dispatchID))
-	current.InferenceRequestID = payload.InferenceRequestId
-	current.Attempt = payload.Attempt
-	current.Outcome = string(payload.Outcome)
-	current.Response = stringValue(payload.Response)
-	current.DurationMillis = payload.DurationMillis
-	current.ExitCode = intPtrValue(payload.ExitCode)
-	current.ErrorClass = stringValue(payload.ErrorClass)
-	current.ProviderSession = interfaces.ProviderSessionMetadataFromGenerated(payload.ProviderSession)
-	current.Diagnostics = interfaces.SafeWorkDiagnosticsFromGenerated(payload.Diagnostics)
-	current.ResponseTime = event.Context.EventTime
-	attempts[payload.InferenceRequestId] = current
-}
-
-func (r *factoryWorldReducer) applyScriptRequest(event factoryapi.FactoryEvent, payload factoryapi.ScriptRequestEventPayload) {
-	if payload.DispatchId == "" || payload.ScriptRequestId == "" {
-		return
-	}
-	requests := r.scriptRequestsForDispatch(payload.DispatchId)
-	current := requests[payload.ScriptRequestId]
-	current.DispatchID = payload.DispatchId
-	current.TransitionID = payload.TransitionId
-	current.ScriptRequestID = payload.ScriptRequestId
-	current.Attempt = payload.Attempt
-	current.Command = payload.Command
-	current.Args = cloneStringSlice(payload.Args)
-	current.RequestTime = event.Context.EventTime
-	requests[payload.ScriptRequestId] = current
-}
-
-func (r *factoryWorldReducer) applyScriptResponse(event factoryapi.FactoryEvent, payload factoryapi.ScriptResponseEventPayload) {
-	if payload.DispatchId == "" || payload.ScriptRequestId == "" {
-		return
-	}
-	responses := r.scriptResponsesForDispatch(payload.DispatchId)
-	current := responses[payload.ScriptRequestId]
-	current.DispatchID = payload.DispatchId
-	current.TransitionID = payload.TransitionId
-	current.ScriptRequestID = payload.ScriptRequestId
-	current.Attempt = payload.Attempt
-	current.Outcome = string(payload.Outcome)
-	current.Stdout = payload.Stdout
-	current.Stderr = payload.Stderr
-	current.DurationMillis = payload.DurationMillis
-	current.ExitCode = intPtrValue(payload.ExitCode)
-	current.FailureType = enumStringValue(payload.FailureType)
-	current.ResponseTime = event.Context.EventTime
-	responses[payload.ScriptRequestId] = current
-}
-
-func (r *factoryWorldReducer) inferenceAttemptsForDispatch(dispatchID string) map[string]interfaces.FactoryWorldInferenceAttempt {
-	attempts := r.stateValue.InferenceAttemptsByDispatchID[dispatchID]
-	if attempts == nil {
-		attempts = make(map[string]interfaces.FactoryWorldInferenceAttempt)
-		r.stateValue.InferenceAttemptsByDispatchID[dispatchID] = attempts
-	}
-	return attempts
-}
-
-func (r *factoryWorldReducer) scriptRequestsForDispatch(dispatchID string) map[string]interfaces.FactoryWorldScriptRequest {
-	requests := r.stateValue.ScriptRequestsByDispatchID[dispatchID]
-	if requests == nil {
-		requests = make(map[string]interfaces.FactoryWorldScriptRequest)
-		r.stateValue.ScriptRequestsByDispatchID[dispatchID] = requests
-	}
-	return requests
-}
-
-func (r *factoryWorldReducer) scriptResponsesForDispatch(dispatchID string) map[string]interfaces.FactoryWorldScriptResponse {
-	responses := r.stateValue.ScriptResponsesByDispatchID[dispatchID]
-	if responses == nil {
-		responses = make(map[string]interfaces.FactoryWorldScriptResponse)
-		r.stateValue.ScriptResponsesByDispatchID[dispatchID] = responses
-	}
-	return responses
-}
-
-func (r *factoryWorldReducer) applyDispatchCompleted(event factoryapi.FactoryEvent, payload factoryapi.DispatchResponseEventPayload) {
-	dispatchID := stringValue(event.Context.DispatchId)
-	if dispatchID == "" {
-		return
-	}
-	dispatch := r.stateValue.ActiveDispatches[dispatchID]
-	delete(r.stateValue.ActiveDispatches, dispatchID)
-
-	workIDs := append([]string(nil), dispatch.WorkItemIDs...)
-	traceIDs := dispatchCompletionTraceIDs(dispatch, event.Context.TraceIds)
-	outputWorkItems, workIDs, traceIDs := r.applyDispatchOutputWork(dispatch, payload, workIDs, traceIDs)
-	r.releaseResourceUnits(dispatch.Resources, payload.OutputResources)
-	completion := r.dispatchCompletionFromResponse(event, payload, dispatchID, dispatch, workIDs, traceIDs, outputWorkItems)
-	r.recordDispatchCompletionState(dispatchID, dispatch, payload, completion)
-}
-
-func dispatchCompletionTraceIDs(dispatch interfaces.FactoryWorldDispatch, eventTraceIDs *[]string) []string {
-	traceIDs := append([]string(nil), dispatch.TraceIDs...)
-	for _, traceID := range sliceValue(eventTraceIDs) {
-		traceIDs = appendUnique(traceIDs, traceID)
-	}
-	return traceIDs
-}
-
-func (r *factoryWorldReducer) applyDispatchOutputWork(
-	dispatch interfaces.FactoryWorldDispatch,
-	payload factoryapi.DispatchResponseEventPayload,
-	workIDs []string,
-	traceIDs []string,
-) ([]interfaces.FactoryWorkItem, []string, []string) {
-	outputWorkItems := make([]interfaces.FactoryWorkItem, 0, len(sliceValue(payload.OutputWork)))
-	for _, work := range sliceValue(payload.OutputWork) {
-		item := r.dispatchOutputWorkItem(dispatch, payload, work)
-		if item.ID == "" {
-			continue
-		}
-		r.stateValue.WorkItemsByID[item.ID] = item
-		workIDs = appendUnique(workIDs, item.ID)
-		traceIDs = appendUnique(traceIDs, item.TraceID)
-		r.addTraceWork(item.TraceID, item.ID)
-		r.addWorkToken(item.ID, item.PlaceID, item)
-		outputWorkItems = append(outputWorkItems, item)
-	}
-	return outputWorkItems, workIDs, traceIDs
-}
-
-func (r *factoryWorldReducer) dispatchOutputWorkItem(
-	dispatch interfaces.FactoryWorldDispatch,
-	payload factoryapi.DispatchResponseEventPayload,
-	work factoryapi.Work,
-) interfaces.FactoryWorkItem {
-	item := factoryWorkItemFromGenerated(work)
-	if item.ID == "" {
-		return interfaces.FactoryWorkItem{}
-	}
-	explicitPlaceID := item.PlaceID
-	previousPlaceID := item.PlaceID
-	if existing, ok := r.stateValue.WorkItemsByID[item.ID]; ok {
-		previousPlaceID = existing.PlaceID
-		item = mergeFactoryWorkItem(existing, item)
-	}
-	if explicitPlaceID == "" {
-		if derivedPlaceID := r.outputPlaceForWork(dispatch.Workstation.ID, payload.Outcome, item.WorkTypeID); derivedPlaceID != "" {
-			item.PlaceID = derivedPlaceID
-		} else if payload.Outcome == factoryapi.WorkOutcomeContinue || payload.Outcome == factoryapi.WorkOutcomeRejected {
-			item.PlaceID = previousPlaceID
-		}
-	}
-	return item
-}
-
-func (r *factoryWorldReducer) dispatchCompletionFromResponse(
-	event factoryapi.FactoryEvent,
-	payload factoryapi.DispatchResponseEventPayload,
-	dispatchID string,
-	dispatch interfaces.FactoryWorldDispatch,
-	workIDs []string,
-	traceIDs []string,
-	outputWorkItems []interfaces.FactoryWorkItem,
-) interfaces.FactoryWorldDispatchCompletion {
-	inputWorkItems := dispatchInputWorkItems(dispatch)
-	latestAttempt := r.latestInferenceAttemptForDispatch(dispatchID)
-	return interfaces.FactoryWorldDispatchCompletion{
-		DispatchID:            dispatchID,
-		TransitionID:          payload.TransitionId,
-		Workstation:           dispatch.Workstation,
-		RunnerID:              dispatch.RunnerID,
-		RunnerSelectionSource: dispatch.RunnerSelectionSource,
-		StartedTick:           dispatch.StartedTick,
-		CompletedTick:         event.Context.Tick,
-		StartedAt:             dispatch.StartedAt,
-		CompletedAt:           event.Context.EventTime,
-		DurationMillis:        int64Value(payload.DurationMillis),
-		Result:                workstationResultFromGenerated(payload),
-		WorkItemIDs:           sortedStrings(workIDs),
-		ConsumedInputs:        interfaces.CloneWorkstationInputs(dispatch.Inputs),
-		InputWorkItems:        sortedWorkItems(inputWorkItems),
-		OutputWorkItems:       sortedWorkItems(outputWorkItems),
-		CurrentChainingTraceID: completedDispatchCurrentChainingTraceID(
-			event.Context.CurrentChainingTraceId,
-			payload.CurrentChainingTraceId,
-			dispatch,
-			inputWorkItems,
-		),
-		PreviousChainingTraceIDs: completedDispatchPreviousChainingTraceIDs(
-			event.Context.PreviousChainingTraceIds,
-			payload.PreviousChainingTraceIds,
-			dispatch,
-			inputWorkItems,
-		),
-		TraceIDs:        interfaces.CanonicalChainingTraceIDs(traceIDs),
-		ProviderSession: latestInferenceProviderSession(latestAttempt),
-		Diagnostics:     latestInferenceDiagnostics(latestAttempt),
-		TerminalWork:    r.terminalWorkForCompletion(payload.Outcome, workIDs),
-	}
-}
-
-func (r *factoryWorldReducer) recordDispatchCompletionState(
-	dispatchID string,
-	dispatch interfaces.FactoryWorldDispatch,
-	payload factoryapi.DispatchResponseEventPayload,
-	completion interfaces.FactoryWorldDispatchCompletion,
-) {
-	r.stateValue.CompletedDispatches = append(r.stateValue.CompletedDispatches, completion)
-	if payload.Outcome == factoryapi.WorkOutcomeFailed {
-		r.stateValue.FailedDispatches = append(r.stateValue.FailedDispatches, completion)
-		r.recordFailedCompletion(completion)
-	}
-	for _, traceID := range completion.TraceIDs {
-		r.addTraceDispatch(traceID, dispatchID)
-	}
-	r.appendProviderSessionRecord(dispatch, payload, completion)
-}
-
-func (r *factoryWorldReducer) appendProviderSessionRecord(
-	dispatch interfaces.FactoryWorldDispatch,
-	payload factoryapi.DispatchResponseEventPayload,
-	completion interfaces.FactoryWorldDispatchCompletion,
-) {
-	if completion.ProviderSession == nil || completion.ProviderSession.ID == "" {
-		return
-	}
-	r.stateValue.ProviderSessions = append(r.stateValue.ProviderSessions, interfaces.FactoryWorldProviderSessionRecord{
-		DispatchID:               completion.DispatchID,
-		TransitionID:             payload.TransitionId,
-		WorkstationName:          dispatch.Workstation.Name,
-		RunnerID:                 completion.RunnerID,
-		RunnerSelectionSource:    completion.RunnerSelectionSource,
-		Outcome:                  string(payload.Outcome),
-		ProviderSession:          *interfaces.CloneProviderSessionMetadata(completion.ProviderSession),
-		WorkItemIDs:              completion.WorkItemIDs,
-		ConsumedInputs:           interfaces.CloneWorkstationInputs(completion.ConsumedInputs),
-		CurrentChainingTraceID:   completion.CurrentChainingTraceID,
-		PreviousChainingTraceIDs: cloneStringSlice(completion.PreviousChainingTraceIDs),
-		TraceIDs:                 cloneStringSlice(completion.TraceIDs),
-		Diagnostics:              interfaces.CloneSafeWorkDiagnostics(completion.Diagnostics),
-		FailureReason:            completion.Result.FailureReason,
-		FailureMessage:           completion.Result.FailureMessage,
-	})
-}
-
-func generatedDispatchRunnerSelection(metadata *factoryapi.DispatchRequestEventMetadata) (string, interfaces.RunnerSelectionSource) {
-	if metadata == nil {
-		return "", ""
-	}
-	return stringValueEnum(metadata.RunnerId), interfaces.RunnerSelectionSource(stringValueEnum(metadata.RunnerSelectionSource))
-}
-
-func dispatchCurrentChainingTraceID(
-	contextCurrent *string,
-	payloadCurrent *string,
-	inputWorkItems []interfaces.FactoryWorkItem,
-) string {
-	if current := stringValue(contextCurrent); current != "" {
-		return current
-	}
-	if current := stringValue(payloadCurrent); current != "" {
-		return current
-	}
-	return interfaces.CurrentChainingTraceIDFromWorkItems(inputWorkItems)
-}
-
-func dispatchPreviousChainingTraceIDs(
-	contextPrevious *[]string,
-	payloadPrevious *[]string,
-	inputWorkItems []interfaces.FactoryWorkItem,
-) []string {
-	if previous := cloneStringSlice(sliceValue(contextPrevious)); len(previous) > 0 {
-		return interfaces.CanonicalChainingTraceIDs(previous)
-	}
-	if previous := cloneStringSlice(sliceValue(payloadPrevious)); len(previous) > 0 {
-		return interfaces.CanonicalChainingTraceIDs(previous)
-	}
-	return interfaces.PreviousChainingTraceIDsFromWorkItems(inputWorkItems)
-}
-
-func completedDispatchCurrentChainingTraceID(
-	contextCurrent *string,
-	payloadCurrent *string,
-	dispatch interfaces.FactoryWorldDispatch,
-	inputWorkItems []interfaces.FactoryWorkItem,
-) string {
-	if current := stringValue(contextCurrent); current != "" {
-		return current
-	}
-	if current := stringValue(payloadCurrent); current != "" {
-		return current
-	}
-	if dispatch.CurrentChainingTraceID != "" {
-		return dispatch.CurrentChainingTraceID
-	}
-	return interfaces.CurrentChainingTraceIDFromWorkItems(inputWorkItems)
-}
-
-func completedDispatchPreviousChainingTraceIDs(
-	contextPrevious *[]string,
-	payloadPrevious *[]string,
-	dispatch interfaces.FactoryWorldDispatch,
-	inputWorkItems []interfaces.FactoryWorkItem,
-) []string {
-	if previous := cloneStringSlice(sliceValue(contextPrevious)); len(previous) > 0 {
-		return interfaces.CanonicalChainingTraceIDs(previous)
-	}
-	if previous := cloneStringSlice(sliceValue(payloadPrevious)); len(previous) > 0 {
-		return interfaces.CanonicalChainingTraceIDs(previous)
-	}
-	if len(dispatch.PreviousChainingTraceIDs) > 0 {
-		return cloneStringSlice(dispatch.PreviousChainingTraceIDs)
-	}
-	return interfaces.PreviousChainingTraceIDsFromWorkItems(inputWorkItems)
-}
-
-func (r *factoryWorldReducer) recordFailedCompletion(completion interfaces.FactoryWorldDispatchCompletion) {
-	if completion.TerminalWork != nil {
-		r.recordFailedWorkDetail(completion, completion.TerminalWork.WorkItem)
-		return
-	}
-	for _, workID := range completion.WorkItemIDs {
-		if item, ok := r.stateValue.WorkItemsByID[workID]; ok {
-			r.recordFailedWorkDetail(completion, item)
-		}
-	}
-}
-
-func dispatchInputWorkItems(
-	dispatch interfaces.FactoryWorldDispatch,
-) []interfaces.FactoryWorkItem {
-	items := make([]interfaces.FactoryWorkItem, 0, len(dispatch.Inputs))
-	for _, input := range dispatch.Inputs {
-		if input.WorkItem == nil || input.WorkItem.ID == "" {
-			continue
-		}
-		items = append(items, *input.WorkItem)
-	}
-	return items
-}
-
-func (r *factoryWorldReducer) latestInferenceAttemptForDispatch(dispatchID string) *interfaces.FactoryWorldInferenceAttempt {
-	attempts := r.stateValue.InferenceAttemptsByDispatchID[dispatchID]
-	if len(attempts) == 0 {
-		return nil
-	}
-	var latest *interfaces.FactoryWorldInferenceAttempt
-	for _, requestID := range sortedMapKeys(attempts) {
-		attempt := attempts[requestID]
-		if latest == nil ||
-			attempt.Attempt > latest.Attempt ||
-			(attempt.Attempt == latest.Attempt && attempt.ResponseTime.After(latest.ResponseTime)) ||
-			(attempt.Attempt == latest.Attempt && attempt.ResponseTime.Equal(latest.ResponseTime) && attempt.InferenceRequestID > latest.InferenceRequestID) {
-			attemptCopy := attempt
-			latest = &attemptCopy
-		}
-	}
-	return latest
-}
-
-func latestInferenceProviderSession(attempt *interfaces.FactoryWorldInferenceAttempt) *interfaces.ProviderSessionMetadata {
-	if attempt == nil {
-		return nil
-	}
-	return interfaces.CloneProviderSessionMetadata(attempt.ProviderSession)
-}
-
-func latestInferenceDiagnostics(attempt *interfaces.FactoryWorldInferenceAttempt) *interfaces.SafeWorkDiagnostics {
-	if attempt == nil {
-		return nil
-	}
-	return interfaces.CloneSafeWorkDiagnostics(attempt.Diagnostics)
-}
-
-func (r *factoryWorldReducer) recordFailedWorkDetail(completion interfaces.FactoryWorldDispatchCompletion, item interfaces.FactoryWorkItem) {
-	if item.ID == "" {
-		return
-	}
-	r.stateValue.WorkItemsByID[item.ID] = item
-	r.stateValue.FailedWorkItemsByID[item.ID] = item
-	delete(r.stateValue.ActiveWorkItemsByID, item.ID)
-	r.addTraceFailed(item.TraceID, item.ID)
-	r.stateValue.FailureDetailsByWorkID[item.ID] = interfaces.FactoryWorldFailureDetail{
-		DispatchID:      completion.DispatchID,
-		TransitionID:    completion.TransitionID,
-		WorkstationName: completion.Workstation.Name,
-		WorkItem:        item,
-		FailureReason:   completion.Result.FailureReason,
-		FailureMessage:  completion.Result.FailureMessage,
-	}
 }
 
 func (r *factoryWorldReducer) applyFactoryStateChange(payload factoryapi.FactoryStateResponseEventPayload) {
@@ -1286,7 +787,7 @@ func factoryWorkItemFromGenerated(work factoryapi.Work) interfaces.FactoryWorkIt
 		CurrentChainingTraceID:   currentChainingTraceID,
 		PreviousChainingTraceIDs: cloneStringSlice(sliceValue(work.PreviousChainingTraceIds)),
 		TraceID:                  traceID,
-		Content:                  workcontent.PartsFromGenerated(work.Content),
+		Content:                  generatedWorkContentToDomain(work.Content),
 		Tags:                     stringMapFromGenerated(work.Tags),
 	}
 }
@@ -1329,6 +830,31 @@ func mergeFactoryWorkItem(existing interfaces.FactoryWorkItem, incoming interfac
 		incoming.Tags = cloneStringMap(existing.Tags)
 	}
 	return incoming
+}
+
+func generatedWorkContentToDomain(content *factoryapi.WorkContent) []interfaces.WorkContentPart {
+	if content == nil || len(*content) == 0 {
+		return nil
+	}
+	parts := make([]interfaces.WorkContentPart, 0, len(*content))
+	for _, part := range *content {
+		textPart, textErr := part.AsWorkTextContentPart()
+		if textErr == nil && textPart.Type == factoryapi.WorkContentPartTypeText {
+			parts = append(parts, interfaces.WorkContentPart{
+				Type: interfaces.WorkContentPartTypeText,
+				Text: textPart.Text,
+			})
+			continue
+		}
+		imagePart, imageErr := part.AsWorkImageContentPart()
+		if imageErr == nil && imagePart.Type == factoryapi.WorkContentPartTypeImage {
+			parts = append(parts, interfaces.WorkContentPart{
+				Type: interfaces.WorkContentPartTypeImage,
+				File: imagePart.File,
+			})
+		}
+	}
+	return parts
 }
 
 func (r *factoryWorldReducer) factoryRelationsFromGenerated(relations *[]factoryapi.Relation, context factoryapi.FactoryEventContext) []interfaces.FactoryRelation {
@@ -1463,155 +989,4 @@ func (r *factoryWorldReducer) topologyPlace(placeID string) (interfaces.FactoryP
 		}
 	}
 	return interfaces.FactoryPlace{}, false
-}
-
-func resourceUnitsFromGenerated(resources *[]factoryapi.Resource) []interfaces.FactoryResourceUnit {
-	if resources == nil {
-		return nil
-	}
-	out := make([]interfaces.FactoryResourceUnit, 0, len(*resources))
-	for _, resource := range *resources {
-		if resource.Name == "" {
-			continue
-		}
-		out = append(out, interfaces.FactoryResourceUnit{ResourceID: resource.Name})
-	}
-	return out
-}
-
-func workstationResultFromGenerated(payload factoryapi.DispatchResponseEventPayload) interfaces.WorkstationResult {
-	return interfaces.WorkstationResult{
-		Outcome:         string(payload.Outcome),
-		Output:          stringValue(payload.Output),
-		Error:           stringValue(payload.Error),
-		Feedback:        stringValue(payload.Feedback),
-		FailureReason:   stringValue(payload.FailureReason),
-		FailureMessage:  stringValue(payload.FailureMessage),
-		ProviderFailure: interfaces.ProviderFailureMetadataFromGenerated(payload.ProviderFailure),
-	}
-}
-
-func placeIDsFromGeneratedIOs(values []factoryapi.WorkstationIO) []string {
-	if len(values) == 0 {
-		return nil
-	}
-	ids := make([]string, 0, len(values))
-	for _, value := range values {
-		ids = append(ids, placeIDFromGeneratedIO(value))
-	}
-	return ids
-}
-
-func placeIDsFromGeneratedIOsPtr(values *[]factoryapi.WorkstationIO) []string {
-	if values == nil {
-		return nil
-	}
-	return placeIDsFromGeneratedIOs(*values)
-}
-
-func placeIDFromGeneratedIO(value factoryapi.WorkstationIO) string {
-	return generatedPlaceID(value.WorkType, value.State)
-}
-
-func generatedPlaceID(workTypeID string, stateValue string) string {
-	if workTypeID == "" || stateValue == "" {
-		return ""
-	}
-	return fmt.Sprintf("%s:%s", workTypeID, stateValue)
-}
-
-func firstRequestID(works *[]factoryapi.Work) string {
-	for _, work := range sliceValue(works) {
-		if requestID := stringValue(work.RequestId); requestID != "" {
-			return requestID
-		}
-	}
-	return ""
-}
-
-func firstString(values *[]string) string {
-	for _, value := range sliceValue(values) {
-		if value != "" {
-			return value
-		}
-	}
-	return ""
-}
-
-func stringMapFromGenerated(values *factoryapi.StringMap) map[string]string {
-	if values == nil {
-		return nil
-	}
-	return cloneStringMap(map[string]string(*values))
-}
-
-func stringValue(value *string) string {
-	if value == nil {
-		return ""
-	}
-	return *value
-}
-
-func stringValueEnum[T ~string](value *T) string {
-	if value == nil {
-		return ""
-	}
-	return string(*value)
-}
-
-func generatedWorkStateName(value *factoryapi.WorkState) string {
-	if value == nil {
-		return ""
-	}
-	return value.Name
-}
-
-func intValue(value *int) int {
-	if value == nil {
-		return 0
-	}
-	return *value
-}
-
-func enumStringValue[T ~string](value *T) string {
-	if value == nil {
-		return ""
-	}
-	return string(*value)
-}
-
-func factoryStateString(value *factoryapi.FactoryState) string {
-	if value == nil {
-		return ""
-	}
-	return string(*value)
-}
-
-func workstationKindString(value *factoryapi.WorkstationKind) string {
-	if value == nil {
-		return ""
-	}
-	return string(*value)
-}
-
-func int64Value(value *int64) int64 {
-	if value == nil {
-		return 0
-	}
-	return *value
-}
-
-func intPtrValue(value *int) *int {
-	if value == nil {
-		return nil
-	}
-	out := *value
-	return &out
-}
-
-func sliceValue[T any](values *[]T) []T {
-	if values == nil {
-		return nil
-	}
-	return *values
 }
